Update CI scripts

+++ ci/build_all_releases.sh
@@ -1,29 +1,41 @@
 #!/bin/bash
-set -ev
+set -v
 
-rm -r ./ci/outputs
+if [ ! -e ".env" ]; then
+    touch .env
+fi 
+if [ -d "./ci/outputs" ]; then
+  rm -r ./ci/outputs
+fi
+mkdir ./ci/outputs
+if [ -d "./build" ]; then
+  rm -r ./build
+fi
 
 # STAGING: Build and zip Rinkeby version
-rm -r ./build
 npm run build:staging
 zip -r ./ci/outputs/staging-rinkeby.zip ./build/*
 
-# PRE-PRODUCTION: Build and zip Rinkeby version
 rm -r ./build
+
+# PRE-PRODUCTION: Build and zip Rinkeby version
 npm run build:preprod
 zip -r ./ci/outputs/preprod-rinkeby.zip ./build/*
 
-# PRE-PRODUCTION: Build and zip Mainnet version
 rm -r ./build
+
+# PRE-PRODUCTION: Build and zip Mainnet version
 npm run build:preprod-mainnet
 zip -r ./ci/outputs/preprod-mainnet.zip ./build/*
 
-# PRODUCTION: Build and zip Rinkeby version
 rm -r ./build
+
+# PRODUCTION: Build and zip Rinkeby version
 npm run build:prod
 zip -r ./ci/outputs/prod-rinkeby.zip ./build/*
 
-# PRODUCTION: Build and zip Mainnet version
 rm -r ./build
+
+# PRODUCTION: Build and zip Mainnet version
 npm run build:prod-mainnet
 zip -r ./ci/outputs/prod-mainnet.zip ./build/*
